requirements.txt
streamlit
openai
tiktoken
python-dotenv
extra-streamlit-components